shuffle=None

+++ tets.py
@@ -207,8 +207,8 @@ def train_model(model, criterion, optimizer, scheduler, dataloaders, num_epochs=
             for i, (img1, img2, target) in enumerate(dataloaders[phase]):
                 if i == epoch_nums[phase]:
                     break
-                img1 = img1.to(device)
-                img2 = img2.to(device)
+                img1 = img1.to(device).half()
+                img2 = img2.to(device).half()
                 target = target.to(device)
                 optimizer.zero_grad()
                 with torch.set_grad_enabled(phase == 'train'):
